add AddFile action
