+++ src/actions/filesActions.js
@@ -1,5 +1,5 @@
 export const FETCH_FILES_SUCCESS = "FETCH_FILES_SUCCESS"
-export const FETCH_FOLDERS_SUCCESS = "FETCH_FOLDERS_SUCCESS"
+export const ADD_FILE_ACTION = "ADD_FILE_ACTION"
 
 // export const ADD_DIRECTORY_ACTION = "ADD_DIRECTORY_ACTION"
 // export const DELETE_DIRECTORY_ACTION = "DELETE_DIRECTORY_ACTION "
@@ -13,7 +13,7 @@ export const makeFetchFilesSuccess = response => ({
   response
 })
 
-export const makeFetchFoldersSuccess = response => ({
-  type: FETCH_FOLDERS_SUCCESS,
+export const addFilesSuccess = response => ({
+  type: ADD_FILE_ACTION,
   response
 })
