+++ src/actions/foldersActions.js
@@ -1,0 +1,6 @@
+export const FETCH_FOLDERS_SUCCESS = "FETCH_FOLDERS_SUCCESS"
+
+export const makeFetchFoldersSuccess = response => ({
+  type: FETCH_FOLDERS_SUCCESS,
+  response
+})
