+++ src/reducers/index.js
@@ -2,8 +2,10 @@ import { combineReducers } from "redux"
 
 import foldersReducers from "./foldersReducer"
 import filesReducers from "./filesReducer"
+import addFilesReducers from "./addFilesReducer"
 
 export default combineReducers({
   folders: foldersReducers,
-  files: filesReducers
+  files: filesReducers,
+  addFiles: addFilesReducers
 })
